Clean up auto-syncing to not resend same geometry
